requirements.txt
fastapi
uvicorn
pydantic
qdrant-client
python-dotenv
google-generativeai
httpx
psutil
